[FEAT / FIX] (Auth) Builder 패턴 추가, 피드백 기반 수정
