+++ src/main/java/com/example/lastproject/domain/auth/dto/request/SignupRequest.java
@@ -3,9 +3,11 @@ package com.example.lastproject.domain.auth.dto.request;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
+@AllArgsConstructor
 public class SignupRequest {
 
     @Email(message = "이메일 형식이 맞지 않습니다.")
