+++ src/main/java/com/example/lastproject/domain/auth/entity/AuthUser.java
@@ -1,6 +1,7 @@
 package com.example.lastproject.domain.auth.entity;
 
 import com.example.lastproject.domain.user.enums.UserRole;
+import lombok.Builder;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -9,16 +10,21 @@ import java.util.Collection;
 import java.util.List;
 
 @Getter
+@Builder
 public class AuthUser {
 
     private final Long userId;
     private final String email;
-    private final Collection<? extends GrantedAuthority> authorities;
+    private final UserRole role;
 
     public AuthUser(Long userId, String email, UserRole role) {
         this.userId = userId;
         this.email = email;
-        this.authorities = List.of(new SimpleGrantedAuthority(role.name()));
+        this.role = role;
+    }
+
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority(role.name()));
     }
 
 }
